+++ src/Model/TeleWallet.java
@@ -9,13 +9,13 @@ public class TeleWallet extends WalletModel{
     }
 
     @Override
-    public void deposit(float amount) {
+    public void deposit(double amount) {
         // Implement deposit logic for TeleWallet
         this.balance += amount;
     }
 
     @Override
-    public void withdraw(int amount) {
+    public void withdraw(double amount) {
         // Implement withdraw logic for TeleWallet
         this.balance -= amount;
     }
